refactor(search, dashboard): extract ui components and hooks
