+++ src/capabilities/search/ui/SearchOverlay.icons.tsx
@@ -1,0 +1,127 @@
+import type { JSX } from 'solid-js'
+import { Show } from 'solid-js'
+import type { SearchResultItemVm } from '~/capabilities/search/ui/search.vm'
+
+export function SearchIcon(): JSX.Element {
+  return (
+    <svg
+      class="h-4 w-4 text-control-foreground"
+      fill="none"
+      stroke="currentColor"
+      viewBox="0 0 24 24"
+      aria-hidden="true"
+    >
+      <path
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="2"
+        d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
+      />
+    </svg>
+  )
+}
+
+export function MatchSourceIcon(props: {
+  readonly source: SearchResultItemVm['matchSource']
+}): JSX.Element {
+  const className = 'h-4 w-4 shrink-0'
+
+  return (
+    <Show
+      when={props.source === 'container'}
+      fallback={
+        <Show
+          when={props.source === 'process'}
+          fallback={
+            <Show
+              when={
+                props.source === 'importer' || props.source === 'bl' || props.source === 'carrier'
+              }
+              fallback={
+                <Show
+                  when={props.source === 'vessel'}
+                  fallback={
+                    <svg
+                      class={`${className} text-tone-info-strong`}
+                      fill="none"
+                      stroke="currentColor"
+                      viewBox="0 0 24 24"
+                      aria-hidden="true"
+                    >
+                      <path
+                        stroke-linecap="round"
+                        stroke-linejoin="round"
+                        stroke-width="2"
+                        d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
+                      />
+                    </svg>
+                  }
+                >
+                  <svg
+                    class={`${className} text-tone-info-fg`}
+                    fill="none"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                    aria-hidden="true"
+                  >
+                    <path
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      stroke-width="2"
+                      d="M3 17l3-3m0 0l4 4 8-8m-8 8V6"
+                    />
+                  </svg>
+                </Show>
+              }
+            >
+              <svg
+                class={`${className} text-text-muted`}
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+                aria-hidden="true"
+              >
+                <path
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"
+                />
+              </svg>
+            </Show>
+          }
+        >
+          <svg
+            class={`${className} text-tone-info-fg`}
+            fill="none"
+            stroke="currentColor"
+            viewBox="0 0 24 24"
+            aria-hidden="true"
+          >
+            <path
+              stroke-linecap="round"
+              stroke-linejoin="round"
+              stroke-width="2"
+              d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
+            />
+          </svg>
+        </Show>
+      }
+    >
+      <svg
+        class={`${className} text-tone-success-fg`}
+        fill="none"
+        stroke="currentColor"
+        viewBox="0 0 24 24"
+        aria-hidden="true"
+      >
+        <path
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          stroke-width="2"
+          d="M5 8h14M5 8a2 2 0 110-4h14a2 2 0 110 4M5 8v10a2 2 0 002 2h10a2 2 0 002-2V8m-9 4h4"
+        />
+      </svg>
+    </Show>
+  )
+}
